pyproject.toml
[project]
name = "lapig-escola"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.0",
    "django>=5.2.4",
    "django-allauth>=65.11.2",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "django-ordered-model>=3.7.4",
    "django-storages>=1.14.6",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.0",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "whitenoise>=6.10.0",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
migrate = "python manage.py makemigrations && python manage.py migrate"
adduser = "python manage.py createsuperuser"

